# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: GreenGenius API
  description: Backend for the GreenGenius AI plant pot.
  version: 1.0.0
paths:
  /v1/devices/claim:
    post:
      tags:
      - devices
      summary: Claim Device
      description: 'Bind a physical pot to the calling user.


        The app reads `claim_code` from the device over an encrypted BLE link, so

        possessing a valid code proves physical proximity. Codes expire after

        `claim_code_ttl_seconds` and are single-use.'
      operationId: claim_device_v1_devices_claim_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/devices:
    get:
      tags:
      - devices
      summary: List Devices
      operationId: list_devices_v1_devices_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceOut'
                title: Response List Devices V1 Devices Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pots:
    get:
      tags:
      - pots
      summary: List Pots
      operationId: list_pots_v1_pots_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PotOut'
                title: Response List Pots V1 Pots Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - pots
      summary: Create Pot
      operationId: create_pot_v1_pots_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PotCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/pots/{pot_id}/water:
    post:
      tags:
      - commands
      summary: Water Pot
      operationId: water_pot_v1_pots__pot_id__water_post
      parameters:
      - name: pot_id
        in: path
        required: true
        schema:
          type: string
          title: Pot Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WaterRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pots/{pot_id}/commands:
    get:
      tags:
      - commands
      summary: List Commands
      operationId: list_commands_v1_pots__pot_id__commands_get
      parameters:
      - name: pot_id
        in: path
        required: true
        schema:
          type: string
          title: Pot Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommandOut'
                title: Response List Commands V1 Pots  Pot Id  Commands Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pots/{pot_id}/identify:
    post:
      tags:
      - identify
      summary: Identify Plant
      description: 'Identify the plant in a photo and bind the species to the pot.


        Multipart rather than base64-in-JSON (what the old endpoint did): base64

        inflates every upload by 33% and forces the whole image into memory as a

        string on both client and server.'
      operationId: identify_plant_v1_pots__pot_id__identify_post
      parameters:
      - name: pot_id
        in: path
        required: true
        schema:
          type: string
          title: Pot Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-dev-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Dev-User
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_identify_plant_v1_pots__pot_id__identify_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Body_identify_plant_v1_pots__pot_id__identify_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_identify_plant_v1_pots__pot_id__identify_post
    ClaimRequest:
      properties:
        claim_code:
          type: string
          maxLength: 32
          minLength: 4
          title: Claim Code
        name:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Name
      type: object
      required:
      - claim_code
      title: ClaimRequest
    ClaimResponse:
      properties:
        device_id:
          type: string
          title: Device Id
        pot_id:
          type: string
          title: Pot Id
        mqtt_username:
          type: string
          title: Mqtt Username
        mqtt_password:
          type: string
          title: Mqtt Password
        mqtt_host:
          type: string
          title: Mqtt Host
        mqtt_port:
          type: integer
          title: Mqtt Port
        mqtt_tls:
          type: boolean
          title: Mqtt Tls
      type: object
      required:
      - device_id
      - pot_id
      - mqtt_username
      - mqtt_password
      - mqtt_host
      - mqtt_port
      - mqtt_tls
      title: ClaimResponse
    CommandOut:
      properties:
        id:
          type: string
          title: Id
        device_id:
          type: string
          title: Device Id
        op:
          type: string
          title: Op
        args:
          additionalProperties: true
          type: object
          title: Args
        state:
          type: string
          title: State
        issued_at:
          type: string
          format: date-time
          title: Issued At
        expires_at:
          type: string
          format: date-time
          title: Expires At
        acked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Acked At
        result:
          anyOf:
          - type: string
          - type: 'null'
          title: Result
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - id
      - device_id
      - op
      - args
      - state
      - issued_at
      - expires_at
      - acked_at
      - result
      - error
      title: CommandOut
    DeviceOut:
      properties:
        id:
          type: string
          title: Id
        model:
          type: string
          title: Model
        fw_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Fw Version
        online:
          type: boolean
          title: Online
        last_seen_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Seen At
      type: object
      required:
      - id
      - model
      - fw_version
      - online
      - last_seen_at
      title: DeviceOut
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IdentifyResponse:
      properties:
        identified:
          type: boolean
          title: Identified
        top:
          anyOf:
          - $ref: '#/components/schemas/IdentifySuggestion'
          - type: 'null'
        suggestions:
          items:
            $ref: '#/components/schemas/IdentifySuggestion'
          type: array
          title: Suggestions
        confidence:
          type: number
          title: Confidence
        species_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Species Id
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
      type: object
      required:
      - identified
      - top
      - suggestions
      - confidence
      title: IdentifyResponse
    IdentifySuggestion:
      properties:
        name:
          type: string
          title: Name
        probability:
          type: number
          title: Probability
        similar_images:
          items:
            type: string
          type: array
          title: Similar Images
      type: object
      required:
      - name
      - probability
      title: IdentifySuggestion
...
    PotCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
      type: object
      required:
      - name
      title: PotCreate
    PotOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        photo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Url
        identify_confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Identify Confidence
        auto_water_enabled:
          type: boolean
          title: Auto Water Enabled
        auto_water_threshold_pct:
          type: number
          title: Auto Water Threshold Pct
        created_at:
          type: string
          format: date-time
          title: Created At
        species:
          anyOf:
          - $ref: '#/components/schemas/SpeciesOut'
          - type: 'null'
      type: object
      required:
      - id
      - name
      - device_id
      - photo_url
      - identify_confidence
      - auto_water_enabled
      - auto_water_threshold_pct
      - created_at
      title: PotOut
...
    SpeciesOut:
      properties:
        id:
          type: string
          title: Id
        scientific_name:
          type: string
          title: Scientific Name
        common_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Common Name
      type: object
      required:
      - id
      - scientific_name
      - common_name
      title: SpeciesOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WaterRequest:
      properties:
        duration_s:
          type: number
          maximum: 30.0
          exclusiveMinimum: 0.0
          title: Duration S
          description: Capped at the firmware interlock limit
      type: object
      required:
      - duration_s
      title: WaterRequest